bot7/requirements.txt
asyncpg==0.30.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
vk-api==11.9.9
